[clean] all structs

+++ include/map/conditions.h
@@ -42,28 +42,28 @@ struct Map_condition {
     /* -- WHO -- */
     // i8     lvl;          /* -1 for anyone    */
     // i8     sex;          /* -1 for anyone    */
-    i16    army;            /* -1 for anyone    */
+    i32    army;            /* -1 for anyone    */
     b32    boss;            /* true for boss    */
     b32    all;             /* true for rout    */
-    i16    unit;            /* -1 for anyone    */
+    i32    unit;            /* -1 for anyone    */
     // i8     unit_class;   /* -1 for anyone    */
     /* -- WHERE -- */
-    // i16    col_min;    /* -1 for anywhere  */
-    // i16    col_max;    /* -1 for anywhere  */
-    // i16    row_min;    /* -1 for anywhere  */
-    // i16    row_max;    /* -1 for anywhere  */
+    // i32    col_min;    /* -1 for anywhere  */
+    // i32    col_max;    /* -1 for anywhere  */
+    // i32    row_min;    /* -1 for anywhere  */
+    // i32    row_max;    /* -1 for anywhere  */
     /* -- WHAT -- */
-    // i16    equippedL;  /* -1 for anything  */
-    // i16    equippedR;  /* -1 for anything  */
+    // i32    equippedL;  /* -1 for anything  */
+    // i32    equippedR;  /* -1 for anything  */
     /* -- WHEN --  */
-    i16    min;        /* -1 for anytime   */
-    i16     at;        /* -1 for never   */
-    i16    max;        /* -1 for anytime   */
+    i32    min;        /* -1 for anytime   */
+    i32     at;        /* -1 for never   */
+    i32    max;        /* -1 for anytime   */
 
     /* --- REWARDS --- */
-    i16    gold;
-    i16    item;
-    i16    scene;
+    i32    gold;
+    i32    item;
+    i32    scene;
 
     /* --- EFFECTS --- */
     b32    win;
